Tasks involving resource handling are all added also into prioritisedTasks set, first use of functor (to tell the set how to sort pointers to Task)
git-svn-id: https://bwapi.googlecode.com/svn/trunk@388 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ bwapi/BWAI/Source/BWAI/Task.cpp
@@ -8,12 +8,12 @@
 namespace BWAI
 {
   //------------------------------- CONSTRUCTOR -------------------------------
-  Task::Task(u8 priority)
+  Task::Task(u16 priority)
   :priority(priority)
   {
   }
   //------------------------------- CONSTRUCTOR -------------------------------
-  Task::Task(Unit* executor, u8 priority)
+  Task::Task(Unit* executor, u16 priority)
   :priority(priority)
   {
     if (executor != NULL)
